feat: atualizar rota de empréstimos para nova página de empréstimos
